OpPoDyn

Documentation for OpPoDyn.

Project Structure

The project is structured as follows

OpPoDyn/
├── assets: contains asses for reference tests
├── docs: contains this documentation
├── OpPoDynTesting: helper package for testing, defines test utilities like reference tests
├── Project.toml
├── src: source code
│   ├── Library: submodule for library, all the models live here
│   └── ...
└── test: test code

At this stage, this project is meant to be used with the main branch from NetworkDynamics. Unfortunately, it also depends on the unregistered subpackage OpPoDynTesting which makes instantiating the environment a bit tricky (because you can neither add NetworkDynamics#main nor OpPoDyntesting#../OpPoDynTesting without it complaining about the other dependency. Thanks to the [sources] block in Project.toml in Julia v1.11, this shouldn'te be a problem anymore.

If you want to use the realse version of Julia v1.10 I suggest to create a new development environment:

julia> pwd() # make sure you're in the right folder
".../.julia/dev/OpPoDyn"

(v1.10) pkg> activate devenv

(devenv) pkg> dev NetworkDynamics

(devenv) pkg> dev ./OpPoDyntesting

(devenv) pkg> dev .

OpPoDyn.ModelMetadataConstructor

OpPoDyn.separate_differential_constraint_eqs

OpPoDyn.small_signal_stability_analysis

OpPoDyn.@attach_metadata!

OpPoDyn.ModelMetadataConstructor — Type

This type wraps the default constructor of a ModelingToolkit.Model and allows to attach metadata to it. Two types of metadata is possible:

a "name" field which will become the "name" field of the system

a named tuple with aribrary fields which will become the "metadata" field of the system

OpPoDyn.separate_differential_constraint_eqs — Function

separate_differential_constraint_eqs(M, p=nothing)

Returns the constraint equations and differential equations indices from an ODEFunction h(x) used in DifferentialEquations.jl. The ODE h must be in Mass Matrix form meaning: M ẋ = h(x), with M diagonal. h should be inplace.

OpPoDyn.small_signal_stability_analysis — Function

small_signal_stability_analysis(h::ODEFunction, eq_point, p = nothing)

Performs a small signal stability analysis according to: [1] "Power System Modelling and Scripting", F. Milano, Chapter 7.2. We find the reduced Jacobian (or State Matrix A_s) and calculate its eigenvalues. If the eigenvalues have positive real parts we classify the grid as unstable.

h: Full DAE system

eq_point: Equilibrium point / fixed point of h. h(eq_point) = 0.0

OpPoDyn.@attach_metadata! — Macro

@attach_metadata! Model metadata

Allows you to attach additonal metadata to a Model which was previously defined using @mtkmodel. The metadata needs to be in the form of a named tuple (; name=..., field1=..., field2=...).

If name is present in the metadata, it will be used as the default name of the system and stripped from the metadata. The rest of the named tuple will be attachde to the ODESystems metadata.

Funding

Development of this project was in part funded by the German Federal Ministry for Economic Affairs and Climate Action as part of the OpPoDyn-Project (Project ID 01258425/1, 2024-2027).